Separate store types.
